Drug recommendation system with decision trees algorithm
In this project, with the aid of cognitive class I have used this classification algorithm to build a model from the historical data of patients, and their response to different medications. I used the trained decision tree to predict the class of an unknown patient, or to find a proper drug for a new patient.
. Machine learning libraries such as pandas, sklearn, numpy etc are used.
. Matplotlib library used for visualization of the drug decision tree.

I have used collected data about a set of patients, all of whom suffered from the same illness. During their course of treatment, each patient responded to one of 5 medications, Drug A, Drug B, Drug c, Drug x and y.

Part of the project is to build a model to find out which drug might be appropriate for a future patient with the same illness. The features of the used dataset are Age, Sex, Blood Pressure, and the Cholesterol of the patients, and the target is the drug that each patient responded to.

It is a sample of multiclass classifier, and we have used the training part of the dataset to build a decision tree, and then used it to predict the class of an unknown patient, or to prescribe a drug to a new patient.