Clustering on vehicle dataset
Hi there, in this project I have successfully implemented hierarchical clustering method with the help of cognitive class AI. The objective here, is to use clustering, to find the most distinctive clusters of vehicles. It will summarize the existing vehicles and help manufacturers to make decision about the supply of new models.

Problem - An automobile manufacturer has developed prototypes for a new vehicle. Before introducing the new model into its range, the manufacturer wants to determine which existing vehicles on the market are most like the prototypes--that is, how vehicles can be grouped, which group is the most similar with the model, and therefore which models they will be competing against.

Tools and technologies used- python libraries such as Scipy, matplotlib, pandas, keras etc
. Agglomerative clustering algorithm using the provided dataset.